Fix bug with exception re-raise
Turns out that if exception is re-raised in `except` block it does not get checked against other `except` blocks for the same `try`. Thus, catch-all `except Exception` does not catch it (as opposed to what was expected of it).

This behaviour would let TypeError in object exposed as api hit the main scope if that error was not caused by inappropriate method call.

This commit fixes that - TypeError that is not caused by inappropriate method call is not getting re-raised.